refactor: flatten structure of listeners
- Renamed `events` package to `listeners`
- Moved hologram sub-packages into a shared `holograms` directory
- Moved singular sub-packages directly under `listeners`
