+++ src/main/java/gg/jos/soulgravesplus/listeners/holograms/DecentHologramsListener.java
@@ -1,4 +1,4 @@
-package gg.jos.soulgravesplus.events.holograms.decentholograms;
+package gg.jos.soulgravesplus.listeners.holograms;
 
 import java.util.UUID;
